Implemented new routes in the storage.js
